# swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: examexercise1
  description: >-

host: examexercise.ru
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
tags:
  - name: Assignement_resource
    description: >-
      A user can put, post, delete and get assignements

paths:
  /assignements:
    get:
      summary: all assignements list
      description: gets the list of all assignements published on the platform
      tags:
        - Assignement_resource
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Assignement'
        '500':
          description: Internal server error
    post:
      summary: user posts new exam
      description: user specifies parameters "taskId", "workerId", "assignementResponse" and a new instance of assignement is created
      tags:
        - Assignement_resource
      parameters:
        - $ref: '#/parameters/Assignement_in_post'
      responses:
        '201':
          description: Created
          schema:
            type: object
            description: the freshly created assignement
        '500':
          description: Internal server error
  '/assignements/{id}':
    put:
      summary: modify assignement resource
      description: user specifies taskId and assignementResponse for the {id}-identified assignement
      tags:
        - Assignement_resource
      parameters:
        - $ref: '#/parameters/Assignement_in_put'
      responses:
        '204':
          description: Success, no content
        '404':
          description: Assignement with id {id} not found
        '500':
          description: Internal server error
    delete:
      summary: deletes assignement resource
      description: the resource specified via {id} is removed from the platform
      tags:
        - Assignement_resource
      parameters:
        - $ref: '#/parameters/Assignement_id_in_path'
      responses:
        '204': Success, no content
        '404':
          description: Assignement with id {id} not found
        '500':
          description: Internal server error

#definitions of different objects
definitions:
  Assignement:
    type: object
    properties:
      assignementId:
        type: string
        description: The assignement's id
      taskId:
        type: string
        description: the taskId of the task the assignement refers to
      workerId:
        type: string
        description: the worker the assignement is assigned to
      assignementResult:
        type: object
        description: An object containing the result for the task
  assignementResult:
    type: object
    properties:
      value:
        type: string
        description: the string that stands as answer to the task
  Assignment_for_post:
  type: object
  properties:
    taskID:
      type: string
      description: the assignment's id
    workerID:
      type: string
      description: the worker's id
      example: 3
    assignmentResult:
      type: 
        - $ref: '#/definitions/assignementResult'
      description: object containing assignement's result
      example: {value: "That is a result"}
  Assignment_for_put:
    type: object
    properties:
      taskId:
        type: string
        description: the assignment's id, eventually modified
      assignementResult:
        type: object
        description: Object containing new assignement result
        example: {value: "That is new result I specify"}
#parameters used in the API communication
parameters:
  Assignement_in_post:
    in: body
    name: assignment
    schema:
      $ref: '#/definitions/Assignment_for_post'
      required:
        - taskID
        - workerID
        - assignmentResult
  Assignement_in_put:
    in: body
    name: assignement
    schema:
      $ref: '#/definitions/Assignment_for_put'
  Assignement_id_in_path:
    name: id
    description: the assignment's id
    in: path
    type: integer
    required: true